backend/requirements.txt
flask
flask_pymongo
flask_cors
pandas
apscheduler
requests
python-dotenv 